parse double

+++ 0x02-functions_nested_loops/6-abs.c
@@ -10,7 +10,7 @@
  */
 int int_abs(int x)
 {
-	return fabs(x);
+	return fabs((double)x);
 }
